+++ tests/datasets/impl/test_eurovoc.py
@@ -18,7 +18,7 @@ from podium.field import Field, MultilabelField
 from podium.storage.resources.large_resource import LargeResource, SCPLargeResource
 from podium.vocab import Vocab
 
-from ..test_eurovoc_dataset import (
+from .test_eurovoc_dataset import (
     crovoc_label_hierarchy,
     documents,
     eurovoc_label_hierarchy,
@@ -292,7 +292,7 @@ def create_mock_dataset(
 
         mappings_path = os.path.join(base_dataset_dir, EuroVocLoader.MAPPING_FILENAME)
 
-        with open("tests/datasets/dataloaders/mock_mapping.xls", mode="rb") as input_file:
+        with open("tests/datasets/impl/mock_mapping.xls", mode="rb") as input_file:
             mappings_content = input_file.read()
         with open(file=mappings_path, mode="wb") as fp:
             fp.write(mappings_content)
