+++ tests/test_vocab.py
