+++ tests/vectorizers/test_vectorizer.py
